clean up gui_utils
